Moved all Roles configurations from Database to Role validator for better DDD (Domain Driven Design)
