+++ src/Database/Database.php
@@ -52,97 +52,6 @@ class Database
         self::PERMISSION_DELETE,
     ];
 
-    // Roles
-    const ROLE_ANY = 'any';
-    const ROLE_GUESTS = 'guests';
-    const ROLE_USERS = 'users';
-    const ROLE_USER = 'user';
-    const ROLE_TEAM = 'team';
-    const ROLE_MEMBER = 'member';
-
-    const ROLES = [
-        self::ROLE_ANY,
-        self::ROLE_GUESTS,
-        self::ROLE_USERS,
-        self::ROLE_USER,
-        self::ROLE_TEAM,
-        self::ROLE_MEMBER,
-    ];
-
-    const ROLE_CONFIGURATION = [
-        Database::ROLE_ANY => [
-            'identifier' => [
-                'allowed' => false,
-                'required' => false,
-            ],
-            'dimension' =>[
-                'allowed' => false,
-                'required' => false,
-            ],
-        ],
-        Database::ROLE_GUESTS => [
-            'identifier' => [
-                'allowed' => false,
-                'required' => false,
-            ],
-            'dimension' =>[
-                'allowed' => false,
-                'required' => false,
-            ],
-        ],
-        Database::ROLE_USERS => [
-            'identifier' => [
-                'allowed' => false,
-                'required' => false,
-            ],
-            'dimension' =>[
-                'allowed' => true,
-                'required' => false,
-                'options' => Database::USER_DIMENSIONS
-            ],
-        ],
-        Database::ROLE_USER => [
-            'identifier' => [
-                'allowed' => true,
-                'required' => true,
-            ],
-            'dimension' =>[
-                'allowed' => true,
-                'required' => false,
-                'options' => Database::USER_DIMENSIONS
-            ],
-        ],
-        Database::ROLE_TEAM => [
-            'identifier' => [
-                'allowed' => true,
-                'required' => true,
-            ],
-            'dimension' =>[
-                'allowed' => true,
-                'required' => false,
-            ],
-        ],
-        Database::ROLE_MEMBER => [
-            'identifier' => [
-                'allowed' => true,
-                'required' => true,
-            ],
-            'dimension' =>[
-                'allowed' => false,
-                'required' => false,
-            ],
-        ],
-    ];
-
-    // Dimensions
-    const DIMENSION_VERIFIED = 'verified';
-    const DIMENSION_UNVERIFIED = 'unverified';
-
-    const USER_DIMENSIONS = [
-        self::DIMENSION_VERIFIED,
-        self::DIMENSION_UNVERIFIED,
-    ];
-
     // Collections
     const METADATA = 'metadata';
